Return error messages as note
In runPseudotimeDE, return an additional column that contains the error message or NA if no error. In tests, check that values are not NA when gene exists, check that there is error message when gene doesn't exist. In addition, use mc.cores=1 to avoid weird bug (on Windows), where the first gene always fails.